pref: divide `game.dynamicStyle`.
